Refactor: 'Main' cleanup
